src/third_party/blink/web_tests/external/wpt/css/css-transitions/CSSPseudoElement-getAnimations.tentative-expected.txt

This is a testharness.js-based test.
FAIL getAnimations sorts simultaneous transitions by name assert_equals: Got pseudo-element target expected "[object CSSPseudoElement]" but got "[object Element]"
Harness: the test ran to completion.
